add ActiveSeries and DeletedSeries doc
